menu added
there is now a start menu, the players do not move until the game is started. a variable stores the start state of the game. every time the player jumps, its velocity is set again, this might need to be changed in the future.
I also added a skull mark when the player is set as dead.